Fixes for iter5 review
